Docs-25 Add Edit page improvements
Refactor layout and localize login and register pages

Replaced headings with PageHeaderComp, adjusted layouts, and localized text in login and register pages. Simplified the grid structure in MainLayout and SubjectDisplayComp. Added editing functionality to AddEditDocPage.

Add support for default editing mode on AddEditDocPage

Updated the AddEditDocPage to support a default editing mode through a boolean URL parameter. Included a condition to show the edit button only if an ID is provided and ensured proper disposal of AuthState.